refactor(native-charts-ng): split package into separate entrypoints
This is a preparation for micro-charts as to avoid that projects
bundle/load unnecessary code.
