+++ projects/native-charts-ng/gauge/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Copyright (c) Siemens 2016 - 2025
+ * SPDX-License-Identifier: MIT
+ */
+export * from './si-nchart-gauge.component';
