+++ projects/native-charts-ng/gauge/si-nchart-gauge.component.html
