+++ projects/native-charts-ng/ng-package.json
@@ -2,6 +2,6 @@
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../dist/@siemens/native-charts-ng",
   "lib": {
-    "entryFile": "src/public-api.ts"
+    "entryFile": "public-api.ts"
   }
 }
