+++ projects/native-charts-ng/utils/index.ts
@@ -2,4 +2,4 @@
  * Copyright (c) Siemens 2016 - 2025
  * SPDX-License-Identifier: MIT
  */
-export * from './si-nchart-gauge/si-nchart-gauge.component';
+export * from './svg-math';
