+++ projects/native-charts-ng/utils/svg-math.ts
